requirements.txt
asgiref==3.5.2
attrs==22.1.0
bcrypt==3.2.2
blessed==1.19.1
boto3==1.24.55
botocore==1.23.42
cached-property==1.5.2
cachetools==5.2.0
cement==2.8.2
certifi==2022.6.15
cffi==1.15.1
charset-normalizer==2.0.12
colorama==0.4.3
cryptography==37.0.4
Django==4.0.6
django-cors-headers==3.13.0
django-friendship==1.9.6
django-storages==1.13.1
dnspython==2.2.1
docker==4.4.4
docker-compose==1.25.5
dockerpty==0.4.1
docopt==0.6.2
filelock==3.8.0
future==0.16.0
gunicorn==20.1.0
httplib2==0.20.4
idna==3.3
jmespath==0.10.0
jsonschema==3.2.0
mysqlclient==2.1.1
oauthlib==3.2.0
paramiko==2.11.0
pathspec==0.9.0
Pillow==9.2.0
protobuf==4.21.5
py3-validate-email==1.0.7
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
PyNaCl==1.5.0
pyparsing==3.0.9
pyrsistent==0.18.1
python-dateutil==2.8.2
pytz==2022.1
PyYAML==5.4.1
requests==2.26.0
requests-oauthlib==1.3.1
rsa==4.9
s3transfer==0.6.0
semantic-version==2.8.5
six==1.14.0
sqlparse==0.4.2
termcolor==1.1.0
texttable==1.6.4
uritemplate==4.1.1
urllib3==1.26.11
wcwidth==0.1.9
websocket-client==0.59.0
